requirements.txt
streamlit
langchain
langchain-openai
langchain-pinecone
langchain-community
pinecone-client
pypdf2
python-dotenv
tiktoken
openai

